requirements.txt
Flask==1.0.3
joblib==0.13.2
matplotlib==3.1.0
opencv-python==4.1.0.25
Pillow==6.0.0
requests==2.22.0